refactor: extract getSignatures
