+++ app/src/main/java/com/neo/marvelCharacters/domain/model/MarvelCharacter.kt
