+++ app/src/main/java/com/neo/marvelCharacters/util/SignatureCompat.kt
@@ -1,0 +1,23 @@
+package com.neo.marvelCharacters.util
+
+import android.annotation.SuppressLint
+import android.content.Context
+import android.content.pm.PackageManager
+import android.content.pm.Signature
+import android.os.Build
+
+object SignatureCompat {
+     fun getSignatures(context: Context): Array<Signature> {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+            context.packageManager.getPackageInfo(
+                context.packageName,
+                PackageManager.GET_SIGNING_CERTIFICATES
+            ).signingInfo.apkContentsSigners
+        } else @SuppressLint("PackageManagerGetSignatures") {
+            context.packageManager.getPackageInfo(
+                context.packageName,
+                PackageManager.GET_SIGNATURES
+            ).signatures
+        }
+    }
+}
